chore: main 페이지 폴더구조 정리
