+++ index.html
@@ -13,7 +13,7 @@
     <meta property="og:url" content="https://script-directors.netlify.app/" />
     <meta property="og:site_name" content="스크립트 디렉터스" />
     <title>스크립트 디렉터스</title>
-    <script type="module" src="/src/scripts/movie_main.js"></script>
+    <script type="module" src="/main/movie_main.js"></script>
   </head>
   <body class="wrap main-wrap">
     <section class="intro-banner-area" aria-labelledby="intro-banner">
