+++ src/main.js
@@ -1,5 +1,6 @@
-import './scripts/components/render/render_header';
-import './scripts/components/render/render_button';
-import './styles/main.css';
+import './scripts/render/render_header';
+import './scripts/render/render_button';
 import './scripts/utils/disable_link';
-import './scripts/components/render/render_footer';
+import './scripts/render/render_footer';
+
+import './styles/main.css';
