--- src/scripts/components/render/reder_main_movies.js
@@ -1,2 +1,0 @@
-import '/src/scripts/api/set_trend_movies';
-import '/src/scripts/api/set_upcoming_movies';
