--- src/scripts/components/render/render_footer.js
@@ -1,7 +1,0 @@
-import { createFooter } from '../ui/create_footer_ui';
-
-function renderFooter() {
-  createFooter();
-}
-
-renderFooter();
