+++ src/scripts/components/ui/create_movie_list.js
