--- src/scripts/movie_main.js
@@ -1,6 +1,0 @@
-import '/src/main.js';
-
-import '/src/scripts/api/set_trend_movies';
-import '/src/scripts/api/set_upcoming_movies';
-
-import '/src/styles/movie_main.css';
