+++ src/scripts/render/reder_main_movies.js
@@ -1,0 +1,2 @@
+import '/main/set_trend_movies';
+import '/main/set_upcoming_movies';
