+++ src/scripts/render/render_footer.js
@@ -1,0 +1,7 @@
+import { createFooter } from '/src/scripts/components/ui/create_footer_ui';
+
+function renderFooter() {
+  createFooter();
+}
+
+renderFooter();
